Apply vendor/inventory patch and verified CSStartSkill layout
Vendor and inventory patch, applied whole, plus one compile fix it needed:
AcquireDefaultItemEx declared a local rollback function that touched the out
parameters, which C# forbids (CS1628). Local lists now back both out params.

Item task packet:
- The trailing block of SC 0x010B is 42 bytes, not 22. Live captures pin it: a
  taskCount=0 packet measures 45 bytes total, which fixes the header at 3 and
  the block at 42, and every other observed length then resolves exactly - a
  money AddStack at 55 = 3+10+42 and a Create at 65 = 3+20+42. We were emitting
  every 0x010B twenty bytes short. Updated the patch's unit test, which asserted
  45 for a single Create by mistaking the zero-task packets for Create packets.
- Restored the chunked inventory resync on the acquire and consume paths. The
  patch dropped it on the theory that 0x010B and 0x0164 suffice; in-game the item
  was then invisible even after an inventory refresh. Its own report concedes the
  client's state gates were never mapped.

Skill packets, from the verified CSStartSkillPacket 0x01A5 report:
- Item caster (type 2) ends in a u64, not a u32. Reading four bytes too few
  shifted everything after the caster block, so the SkillObject header came out
  of the middle of that value - the real cause of the coin purse crash, exactly
  where the report predicted the fault would be.
- Rewrote the SkillObject table to the verified types 1..30. Ours came from a
  3.0+ client and was shifted by one from type 6 onward, so every object above 5
  decoded with the wrong layout. Type 28 (sit/lie down) is now read rather than
  ignored, type 8 is a fixed 48-byte block, and type 14 always carries 50 values.
- inputDirection is written after every payload including type 0.

+++ AAEmu.Game/Models/Game/Items/Actions/MoneyChangeBank.cs
@@ -4,9 +4,9 @@ namespace AAEmu.Game.Models.Game.Items.Actions;
 
 public class MoneyChangeBank : ItemTask
 {
-    private readonly int _amount;
+    private readonly long _amount;
 
-    public MoneyChangeBank(int amount)
+    public MoneyChangeBank(long amount)
     {
         _type = ItemAction.ChangeBankMoneyAmount; // 2
         _amount = amount;
@@ -15,7 +15,7 @@ public class MoneyChangeBank : ItemTask
     public override PacketStream Write(PacketStream stream)
     {
         base.Write(stream);
-        stream.Write(_amount); // amount
+        stream.Write(_amount); // i64 in target 1.8.1.0 serializer
         return stream;
     }
 }
